ef_copert5

Functions to calculate vehicle speed-based hot exhaust PM2.5 and PM10 emission factors based on COPERT5.0 methods.

Usage

ef_copert5_exh_pm(
  veh.spd = NULL,
  veh.type = NULL,
  veh.wt = NULL,
  em.type = "pm",
  route.def = NULL,
  fuel.type = NULL,
  fuel.corr = TRUE,
  em.source = "exhaust",
  euro.class = NULL,
  eng.load = NULL,
  route.slope = NULL,
  ...
)

ef_copert5_exh_pm2.5(...)

ef_copert5_exh_pm10(...)

Source

Average speed curves are based on COPERT 5 methods as documented in:

Dan Wakeling (Ricardo Energy & Environment). Road vehicle emission factors for PM based on COPERT v5.0 (2017): speed-emission factor equations. NAEI Ref: ED62553001 https://naei.beis.gov.uk/data/ef-transport

Fuel corrections are based on:

TRL Emission Factors - scaling factors for mileage and improvements in fuel quality https://www.gov.uk/government/publications/road-vehicle-emission-factors-2009

Arguments

veh.spd

average vehicle speed, in km/h.

veh.type

vehicle type, currently only coded for urban buses, see Note.

veh.wt

vehicle weight, in km.

em.type

type of emission, currently PM, PM2.5 or PM10, see Note.

route.def

route description, optional for ef_copert5 functions.

fuel.type

fuel used by the vehicle, diesel, etc.

fuel.corr

(logical) apply fuel correction, see Note.

em.source

emission source: currently, exhaust only.

euro.class

EURO classification: PRE, I, II, III, IV, V+ERG, V+SCR or VI only.

eng.load

engine load, as proportion: 0. 0.5 or 1 only.

route.slope

route slope, as proportion: -0.06, -0.04, -0.02, 0, 0.02, 0.04 or 0.06 only.

...

other arguments, currently ignored

Note

The embrs::ef_copert5 functions are currently only coded for urban buses, veh.type="bus".

Strictly, COPERT PM exhaust emission factors are PM, but because vehicle exhaust PM is fine it is typically assumed that:

PM10 = PM2.5 = PM for exhaust emissions from most conventional diesel and gasoline vehicles.

Fuel Corrections are applied to account for improvements in emissions associated with improvements in fuel quality,